requirements.txt
pandas
numpy
requests
python-dotenv
matplotlib
seaborn
mplfinance 